more tweaks working

+++ overseer/overseer.py
@@ -289,17 +289,18 @@ class Overseer:
         # console.log(f"{self.p} expected len {segment_length:.04f} but found {midi.length:.04f} and calcd {mido.tick2second(total_time_t, 220, mido.bpm2tempo(file_bpm))}")
 
         # scale velocities
-        new_pm = PrettyMIDI(new_file_path)
-        os.remove(new_file_path)
-        [[v_min_o, v_max_o], v_hist_o] = um.get_velocities(new_pm)
-        scaled_midi = um.scale_vels(new_pm, self.v_scale)
-        [[v_min_i, v_max_i], v_hist_i] = um.get_velocities(scaled_midi)
-        scaled_midi.write(new_file_path)
-        console.log(
-            f"{self.p} scaled by factor {self.v_scale} ({v_min_o}, {v_max_o}) -> ({v_min_i}, {v_max_i})\n{v_hist_o} -> {v_hist_i}"
-        )
+        if self.v_scale != 1.0 :
+            new_pm = PrettyMIDI(new_file_path)
+            os.remove(new_file_path)
+            [[v_min_o, v_max_o], v_hist_o] = um.get_velocities(new_pm)
+            scaled_midi = um.scale_vels(new_pm, self.v_scale)
+            [[v_min_i, v_max_i], v_hist_i] = um.get_velocities(scaled_midi)
+            console.log(
+                f"{self.p} scaled by factor {self.v_scale} ({v_min_o}, {v_max_o}) -> ({v_min_i}, {v_max_i})\n{v_hist_o} -> {v_hist_i}"
+            )
+            scaled_midi.write(new_file_path)
 
-        old_path = os.path.join(self.plot_dir, os.path.basename(midi_file_path))
+        # old_path = os.path.join(self.plot_dir, os.path.basename(midi_file_path))
         old_pr = PrettyMIDI(midi_file_path).get_piano_roll()
         new_pr = PrettyMIDI(new_file_path).get_piano_roll()
         plot_path = os.path.join(self.plot_dir, f"loop {self.iter}.png")
